% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Parallel Manhattan distance}
\usage{
distance(x, loop = c("standard", "colwise", "diagwise"))
}
\arguments{
\item{x}{An integer matrix. Missing values are not allowed.}

\item{loop}{Possible values are "standard", "colwise", and "diagwise".
With "standard", the code is run using two loops (for the two rows being 
compared). With "colwise" and "diagwise", the code is run using a single 
loop on an index then used to compute the row indices, in a 
"col-wise" and "diag-wise" way, respectively.}
}
\value{
An integer vector of distances for all the pairwise combinations.
Note that \code{loop = "diagwise"} will return values in a different order.
}
\description{
Parallel Manhattan distance using C and OpenMP.
This is a basic distance computation for integer values only.
It has been written in order to compare different looping strategies
and their impact on computation time.
This function is experimental and has a limited scope. It is not meant 
to be used for a serious purpose.
}
