% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_procs}
\alias{num_procs}
\title{Number of Processors}
\usage{
num_procs()
}
\description{
Returns the total number of physical+logical processors, as
detected by omp_get_max_threads().  Will  probably agree with 
\code{parallel::detectCores()}.     
Code written by Drew Schmidt.
}
